+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 2.2.0 - 2019-03-08
+### Add
+- additional validation
+- dynamic password progress indicator
+- show/hide password
+
 ## 2.1.2 - 2019-03-04
 ### Fixed
 - forgot to run build, rebuild package with new changes